Split helper_functions.py into multiple files within utils directory
