+++ backend/src/nasdaq_listings.py
@@ -2,7 +2,8 @@ import requests
 import pandas as pd
 import json
 from requests.exceptions import Timeout
-from helper_functions import print_status, create_outfile
+from utils.logging import print_status
+from utils.outfiles import create_outfile
 
 # print header message to terminal
 process_name = "NASDAQ Listings"
