+++ backend/src/relative_strengths.py
@@ -1,6 +1,8 @@
 import yfinance as yf
 import pandas as pd
-from helper_functions import *
+from utils.logging import *
+from utils.outfiles import *
+from utils.calculations import *
 
 # minimum RS required to pass this screen
 min_rs = 90
