+++ backend/src/trend.py
@@ -1,4 +1,3 @@
-from helper_functions import *
 from selenium import webdriver
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.by import By
@@ -10,9 +9,13 @@ from multiprocessing.pool import ThreadPool
 from tqdm import tqdm
 import requests
 from lxml import html
+from utils.logging import *
+from utils.outfiles import *
+from utils.calculations import *
+from utils.scraping import *
 
 # constants
-threads = 5  # number of concurrent Selenium browser instances to fetch data
+threads = 10  # number of concurrent Selenium browser instances to fetch data
 timeout = 30
 sma_10_xpath = "/html/body/div[3]/div[4]/div[2]/div[2]/div/section/div/div[6]/div[2]/div[2]/table/tbody/tr[3]/td[2]"
 sma_20_xpath = "/html/body/div[3]/div[4]/div[2]/div[2]/div/section/div/div[6]/div[2]/div[2]/table/tbody/tr[5]/td[2]"
